cleanups and improvs

+++ backend/app/app/utils/filestorage.py
@@ -48,12 +48,16 @@ class LocalFile:
         return LocalFile(meta["name"], meta["type"], file_id)
 
 
-async def store_file_into_local_cache(*, reg_ticket_txid, file_bytes, extra_suffix: str = "") -> str:
+async def store_file_into_local_cache(*, reg_ticket_txid, file_bytes,
+                                      overwrite: bool = False, extra_suffix: str = "") -> str:
     cached_result_file = \
         f"{settings.FILE_STORAGE}/{settings.FILE_STORAGE_FOR_RESULTS_SUFFIX}/{reg_ticket_txid}{extra_suffix}"
     try:
         if not os.path.exists(f"{settings.FILE_STORAGE}/{settings.FILE_STORAGE_FOR_RESULTS_SUFFIX}"):
             os.makedirs(f"{settings.FILE_STORAGE}/{settings.FILE_STORAGE_FOR_RESULTS_SUFFIX}")
+
+        if not overwrite and os.path.exists(cached_result_file):
+            return cached_result_file
 
         with open(cached_result_file, 'wb') as f:
             f.write(file_bytes)
@@ -75,7 +79,6 @@ async def search_file_in_local_cache(*, reg_ticket_txid, extra_suffix: str = "")
 async def search_processed_file(*, db, task_from_db, update_task_in_db_func,
                                 task_done, service: wn.WalletNodeService) -> bytes:
     file_bytes = await search_file_in_local_cache(reg_ticket_txid=task_from_db.reg_ticket_txid)
-    not_locally_cached = not file_bytes
 
     if not file_bytes and task_done:
         file_bytes = await wn.get_file_from_pastel(reg_ticket_txid=task_from_db.reg_ticket_txid, wn_service=service)
@@ -87,20 +90,18 @@ async def search_processed_file(*, db, task_from_db, update_task_in_db_func,
         raise PastelAPIException(f"Dupe detection data is not found")
 
     # cache file in local storage and IPFS
-    if not_locally_cached:
-        cached_result_file = await store_file_into_local_cache(reg_ticket_txid=task_from_db.reg_ticket_txid,
-                                                               file_bytes=file_bytes)
-        if cached_result_file and not task_from_db.stored_file_ipfs_link:
-            stored_file_ipfs_link = await store_file_to_ipfs(cached_result_file)
-            if stored_file_ipfs_link:
-                upd = {"stored_file_ipfs_link": stored_file_ipfs_link, "updated_at": datetime.utcnow()}
-                update_task_in_db_func(db, db_obj=task_from_db, obj_in=upd)
+    cached_result_file = await store_file_into_local_cache(reg_ticket_txid=task_from_db.reg_ticket_txid,
+                                                           file_bytes=file_bytes)
+    if cached_result_file and not task_from_db.stored_file_ipfs_link:
+        stored_file_ipfs_link = await store_file_to_ipfs(cached_result_file)
+        if stored_file_ipfs_link:
+            upd = {"stored_file_ipfs_link": stored_file_ipfs_link, "updated_at": datetime.utcnow()}
+            update_task_in_db_func(db, db_obj=task_from_db, obj_in=upd)
     return file_bytes
 
 
 async def search_nft_dd_result(*, db, task_from_db, update_task_in_db_func) -> bytes:
     dd_bytes = await search_file_in_local_cache(reg_ticket_txid=task_from_db.reg_ticket_txid, extra_suffix=".dd")
-    not_locally_cached = not dd_bytes
 
     if not dd_bytes:
         dd_bytes = await wn.get_nft_dd_result_from_pastel(reg_ticket_txid=task_from_db.reg_ticket_txid)
@@ -112,40 +113,13 @@ async def search_nft_dd_result(*, db, task_from_db, update_task_in_db_func) -> b
         raise PastelAPIException(f"Dupe detection data is not found")
 
     # cache file in local storage and IPFS
-    if not_locally_cached:
-        cached_dd_file = await store_file_into_local_cache(reg_ticket_txid=task_from_db.reg_ticket_txid,
-                                                           file_bytes=dd_bytes,
-                                                           extra_suffix=".dd")
-        if cached_dd_file and not task_from_db.nft_dd_file_ipfs_link:
-            nft_dd_file_ipfs_link = await store_file_to_ipfs(cached_dd_file)
-            if nft_dd_file_ipfs_link:
-                upd = {"nft_dd_file_ipfs_link": nft_dd_file_ipfs_link, "updated_at": datetime.utcnow()}
-                update_task_in_db_func(db, db_obj=task_from_db, obj_in=upd)
+    cached_dd_file = await store_file_into_local_cache(reg_ticket_txid=task_from_db.reg_ticket_txid,
+                                                       file_bytes=dd_bytes,
+                                                       extra_suffix=".dd")
+    if cached_dd_file and not task_from_db.nft_dd_file_ipfs_link:
+        nft_dd_file_ipfs_link = await store_file_to_ipfs(cached_dd_file)
+        if nft_dd_file_ipfs_link:
+            upd = {"nft_dd_file_ipfs_link": nft_dd_file_ipfs_link, "updated_at": datetime.utcnow()}
+            update_task_in_db_func(db, db_obj=task_from_db, obj_in=upd)
 
     return dd_bytes
-
-
-async def search_processed_file(*, db, task_from_db, update_task_in_db_func,
-                                task_done, service: wn.WalletNodeService) -> bytes:
-    file_bytes = await search_file_in_local_cache(reg_ticket_txid=task_from_db.reg_ticket_txid)
-    not_locally_cached = not file_bytes
-
-    if not file_bytes and task_done:
-        file_bytes = await wn.get_file_from_pastel(reg_ticket_txid=task_from_db.reg_ticket_txid, wn_service=service)
-
-    if not file_bytes and task_from_db.stored_file_ipfs_link:
-        file_bytes = await read_file_from_ipfs(task_from_db.stored_file_ipfs_link)
-
-    if not file_bytes:
-        raise PastelAPIException(f"Processed file is not found")
-
-    # cache file in local storage and IPFS
-    if not_locally_cached:
-        cached_result_file = await store_file_into_local_cache(reg_ticket_txid=task_from_db.reg_ticket_txid,
-                                                               file_bytes=file_bytes)
-        if cached_result_file and not task_from_db.stored_file_ipfs_link:
-            stored_file_ipfs_link = await store_file_to_ipfs(cached_result_file)
-            if stored_file_ipfs_link:
-                upd = {"stored_file_ipfs_link": stored_file_ipfs_link, "updated_at": datetime.utcnow()}
-                update_task_in_db_func(db, db_obj=task_from_db, obj_in=upd)
-    return file_bytes
